console: Don't unbind console evtchn on suspend
The console evtchn is bound by the domain builder, if we unbind the
evtchn before suspend it may be impossible to rebind it in case of a
canceled suspend. The suspend resume logic was rewritten to accommodate
this changes: the suspend masks the evtchn and the resume unmask it
again, the bind_evtchn is not necessary anymore because the handler is
set the first time we bind the evtchn (on init).