xenbus: Handle canceled suspend
The xenstore evtchn is bound by the domain builder, if we unbind the
evtchn before suspend it may be impossible to rebind it in case of a
canceled suspend. The suspend resume logic was rewritten to accommodate
this change.

In case of a canceled suspend the watches shouldn't be restored.