netfront: File descriptor operation and interface
File descriptor operation is execued in the block of HAVE_LIBC
definition. I added CONFIG_SELECT_POLL option to handle the file
descriptor without HAVE_LIBC option.
netfront_dev structure is defined in netfront.c and we need an interface
to access the member of the structure. I add a function to access the
file descriptor which is the member of netfront_dev.

Signed-off-by: Kenichi Yasukata <kenichi.yasukata@neclab.eu>