lib: ASSERT and BUG_ON macros for debug builds
The macros ASSERT() and BUG_ON() are useful for developing
because they are intended to check for known conditions that
would lead to a bad system state. If such a condition is
fulfilled, the system crashes gracefully and prints the
condition to the console. Because these macros add CPU
instructions to do the condition checking, it introduces
overhead.

With this commit, these macros are replaced by C statements
with no effect whenever debugging is disabled. This way, we
save CPU cycles on such builds.