reduce file size of linked binary
Explanation: ld by default aligns sections to 2M pages for the mini-os
binary. This can be disabled by using either --nmagic or --omagic. The
difference is that omagic additionaly sets .text and .data to be
writable, potentially allowing further space-saving consolidation. At
least at this point in time, the two option make no difference for
mini-os, probably becaue of the enforced alignment specified in the
custom linker file. But who knows, maybe it might in the future.